+++ Makefile
@@ -31,9 +31,9 @@ test:
 
 testsuite:
 	cd tools/Test-CentOS7; ./run_test_in_docker.sh file
-	cd tools/Test-CentOS7; ./run_test_in_docker.sh s3
-	cd tools/Test-CentOS7; ./run_test_in_docker.sh minio
-	cd tools/Test-CentOS7; ./run_test_in_docker.sh minioCs
+	cd tools/Test-CentOS7; ./run_test_in_docker.sh s3aws
+	cd tools/Test-CentOS7; ./run_test_in_docker.sh s3minio
+	cd tools/Test-CentOS7; ./run_test_in_docker.sh s3minioCs
 
 tarball: man
 	mkdir -p $(BUILD)/docs
