+++ circle.yml
@@ -14,6 +14,6 @@ test:
   override:
     - make test
     - cd tools/Test-CentOS7; ./run_test_in_docker.sh file
-    - cd tools/Test-CentOS7; ./run_test_in_docker.sh s3
-    - cd tools/Test-CentOS7; ./run_test_in_docker.sh minio
-    - cd tools/Test-CentOS7; ./run_test_in_docker.sh minioCs
+    - cd tools/Test-CentOS7; ./run_test_in_docker.sh s3aws
+    - cd tools/Test-CentOS7; ./run_test_in_docker.sh s3minio
+    - cd tools/Test-CentOS7; ./run_test_in_docker.sh s3minioCs
