+++ docs/config-example.yml
@@ -38,10 +38,10 @@
 # Concurrent threads, default: calculateted based on cores
 #jobs: 4 
 
-# Backup destination (file|minios3|awss3|minioCs)
+# Backup destination (file|s3minio|s3aws|s3minioCs)
 #backup_to: file 
 
-# WAL destination (file|minios3|awss3|minioCs)
+# WAL destination (file|s3minio|s3aws|s3minioCs)
 #archive_to: file 
 
 # S3 Endpoint IP and Port
