+++ storage/backends/s3minioCs/s3minioCs.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package minioCs
+package s3minioCs
 
 import (
 	"bytes"
